Do not write the company_id on the line
The field company_id is a related towards the move line.
If we write in credit_control_line.company_id, the ORM will
do a write on the company_id field of account_move_line.

It is useless anyway and if the module account_constraints is installed, it
will block the creation of credit_control_line records.